+++ internal/sync/abi/gps_verifier_abi.json
