+++ internal/sync/abi/starknet_abi.json
